Added pygame-based version of sense_hat_show_ip.py
This version has a couple of limitations: The scrolling text finishes
scrolling before exiting after exit requested, and it also needs to
run within an X session (needed by pygame). A Pi is normally running
X on its console even if no monitor is plugged in to see it; however
were X to be disabled it would break the program which is why I
wrote a separate version of the program using threading, see that
Git branch for that version.